Improve page-in-page on iOS
CreateViewController was causing a new Application object to be
instantiated which then got set as the Current application. This caused
global styling to break as well as breaking custom applications that
store expected state. It also now uses the exposed Platform APIs and
attempts to reduce extra UIViews from being created.